requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
sqlalchemy
google-genai
python-multipart